pyproject.toml
[project]
name = "fasthtml-retro-os"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "python-fasthtml>=0.12.21",
    "ruff>=0.12.3",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
